changed order creation to metadata based system

+++ db/migrate/20200508102319_change_amount_to_be_decimal_in_billings.rb
@@ -1,6 +1,6 @@
 class ChangeAmountToBeDecimalInBillings < ActiveRecord::Migration[6.0]
   def self.up
-  	change_column :billings, :amount, :decimal, precision: 10, scale: 6
+  	change_column :billings, :amount, :decimal, precision: 16, scale: 6
   end
 
   def self.down
